+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased changes](https://github.com/natemcmaster/CommandLineUtils/compare/v5.0.1...main)
+
+### Features
+* [@Chris-Wolfgang]: Add support for keyed dependency injection via `[FromKeyedServices]` attribute ([#560])
+
+[#560]: https://github.com/natemcmaster/CommandLineUtils/pull/560
+
 ## [v5.0.1](https://github.com/natemcmaster/CommandLineUtils/compare/v5.0.0...v5.0.1)
 
 ### Features